requirements.txt
web3==5.30.0
python-dotenv==0.20.0